+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `CSVRecordReader` implements `IEnumerable<string>`.
 
 ## [1.3.1] - 2022-02-11
 ### Added
@@ -64,7 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Read and write csv form feature. Support table cells contain commas and double quotes.
 
-[Unreleased]: https://github.com/Mr-sB/TinyCSV/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/Mr-sB/TinyCSV/compare/v1.3.1...HEAD
 [1.3.1]: https://github.com/Mr-sB/TinyCSV/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/Mr-sB/TinyCSV/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Mr-sB/TinyCSV/compare/v1.1.0...v1.2.0